chore: rename package to com.expediagroup.openworld.sdk
