+++ openapi/src/main/kotlin/com/expediagroup/openworld/sdk/generators/openapi/MustacheHelpers.kt
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.expediagroup.sdk.generators.openapi
+package com.expediagroup.openworld.sdk.generators.openapi
 
 import com.samskivert.mustache.Mustache
